Praise our robot Overlords, committing code is my purpose
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36558 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/plus4/plus4cart.c
@@ -398,7 +398,9 @@ int cartridge_attach_image(int type, const char *filename)
             if ((type & 0xff00) == CARTRIDGE_PLUS4_DETECT) {
                 return cart_load_generic(type, filename);
             } else {
-                log_error(LOG_DEFAULT, "cartridge_attach_image: unsupported type (%04x)", type);
+                log_error(LOG_DEFAULT,
+                        "cartridge_attach_image: unsupported type (%04x)",
+                        (unsigned int)type);
             }
             break;
     }
